lss_gesture_arm/requirements.txt
lynxmotion-lss>=0.1.0
mediapipe>=0.10.0
opencv-python>=4.9.0
numpy>=1.26.0
